refactor: move injectCatppuccin to module
