+++ src/scripts/script-utils.js
@@ -6,6 +6,27 @@ export function injectCSS(css) {
   document.head.appendChild(link);
 }
 
+export function injectCatppuccin(flavour, accent) {
+  console.log("injecting catppuccin theme");
+  fetch(browser.runtime.getURL("themes/catppuccin.json"))
+    .then((response) => response.json())
+    .then((palette) => {
+      let style = document.createElement("style");
+      style.classList.add("schooltape");
+      let cssText = "";
+      for (let color in palette[flavour]["colors"]) {
+        let c = palette[flavour]["colors"][color];
+        let hsl = `${c.hsl.h} ${c.hsl.s * 100}% ${c.hsl.l * 100}%`;
+        cssText += `:root { --ctp-${color}: ${hsl}; }\n`;
+      }
+      let a = palette[flavour]["colors"][accent].hsl;
+      cssText += `:root { --ctp-accent: ${`${a.h} ${a.s * 100}% ${a.l * 100}%`}; }\n`;
+      style.textContent = cssText;
+      document.head.appendChild(style);
+      console.log(style);
+    });
+}
+
 // export function injectPlugin(pluginName, loadTime) {
 //   // inject plugins
 //   let xhr = new XMLHttpRequest();
